% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_at_intervals.R
\name{app_at_intervals}
\alias{app_at_intervals}
\title{Launch \code{\link{at_intervals}} shiny application.}
\usage{
app_at_intervals(...)
}
\arguments{
\item{...}{Arguments to \code{\link[shiny]{runApp}}.}
}
\description{
This is a wrapper function for launching a shiny application example of the
\code{\link{at_intervals}} function.
}
\examples{
\dontrun{
launch_application()
}

}

