% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_intervals.R
\name{at_intervals}
\alias{at_intervals}
\title{Values of arbitrary time series at regular intervals.}
\usage{
at_intervals(x, ..., na.rm = TRUE)
}
\arguments{
\item{x}{A data frame}

\item{...}{Arguments passed to \code{seq} method}

\item{na.rm}{A logical scalar.  Defaults to TRUE.}
}
\value{
A data frame
}
\description{
Time series are assumed to be structured as a data frame with one column
being time stamps and the remaining columns being features that are observed.
In general, features are observed irregularly and asynchronously.  The
purpose of this function is derive a regular sequenced, synchronised time
series.
}
\details{
The function is designed to be used with the \link[dplyr]{dplyr} family.
}
\examples{
 x <- tibble::tribble(
   ~ts,                   ~x, ~y, ~z,
 #|----------------------|---|---|---|
   "2016-01-01 00:00:00", NA, 1,  1,
   "2016-01-01 00:05:00", 2,  NA, NA,
   "2016-01-01 00:10:00", 3,  3,  NA
   )

 x <- dplyr::mutate(x, ts = lubridate::ymd_hms(ts))

 at_intervals(x, by = "2 min", na.rm = FALSE)

}
\seealso{
\code{\link[base]{seq}}
}

